backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
psycopg2-binary==2.9.7
cryptography==41.0.4
qrcode[pil]==7.4.2
reportlab==4.0.4
python-dotenv==1.0.0
Pillow==10.0.1